requirements.txt
fastapi==0.109.0
python-dotenv==1.0.1
uvicorn==0.27.0.post1
flask-cors==3.0.10
boto3==1.34.32
requests==2.31.0
openai==1.10.0
orjson==3.9.12
python-multipart==0.0.6